+++ Sources/KeyboardKit/Extensions/String+Sentence.swift
@@ -1,5 +1,5 @@
 //
-//  SentenceAnalyzer.swift
+//  String+Sentence.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-12-28.
@@ -8,56 +8,6 @@
 
 import Foundation
 
-/**
- This protocol can be implemented by any type that should be
- able to analyze sentence information for strings.
-
- Implementing the protocol extends types with functions that
- use public `String` extensions with the same names.
- 
- While you can use this protocol, the main reason to have it
- is to expose these extensions to DocC.
- */
-public protocol SentenceAnalyzer {}
-
-public extension SentenceAnalyzer {
-
-    /**
-     Check whether or not the last character in the provided
-     string is a sentence delimiter.
-     */
-    func hasSentenceDelimiterSuffix(in string: String) -> Bool {
-        string.hasSentenceDelimiterSuffix
-    }
-
-    /**
-     Check whether or not the last sentence in the string is
-     ended, with or without trailing whitespace.
-     */
-    func isLastSentenceEnded(in string: String) -> Bool {
-        string.isLastSentenceEnded
-    }
-
-    /**
-     Check whether or not the last sentence in the string is
-     ended with trailing whitespace.
-     */
-    func isLastSentenceEndedWithTrailingWhitespace(in string: String) -> Bool {
-        string.isLastSentenceEndedWithTrailingWhitespace
-    }
-
-    /**
-     Get the content of the last sentence, if any. Note that
-     it will not contain the sentence delimiter.
-     */
-    func lastSentence(in string: String) -> String? {
-        string.lastSentence
-    }
-}
-
-
-// MARK: - String
-
 public extension String {
 
     /**
@@ -65,7 +15,8 @@ public extension String {
      is a sentence delimiter.
      */
     var hasSentenceDelimiterSuffix: Bool {
-        String(last ?? Character("")).isSentenceDelimiter
+        guard let last else { return false }
+        return String(last).isSentenceDelimiter
     }
 
     /**